create diane-conductor.py
Diane's version of brain functionality. To be merged somehow with Jeff's version.

What I have so far is mainly some infrastructure, a routine that tells which day it is plus which portion (sunrise med, daytime, sunset med, nighttime) we are on, and has a lot of testing abilities built in, including running quickly through a day's, or several day's, programming.